Product model 분리
