+++ src/main/java/com/kh/devrun/admin/model/dao/AdminDao.java
@@ -4,25 +4,15 @@ import java.util.List;
 import java.util.Map;
 
 import com.kh.devrun.category.model.vo.ProductChildCategory;
-import com.kh.devrun.product.Product;
-import com.kh.devrun.product.ProductCategory;
-import com.kh.devrun.product.ProductDetail;
-import com.kh.devrun.product.ProductExtends;
+import com.kh.devrun.product.model.vo.Product;
+import com.kh.devrun.product.model.vo.ProductCategory;
+import com.kh.devrun.product.model.vo.ProductDetail;
+import com.kh.devrun.product.model.vo.ProductExtends;
 import com.kh.devrun.promotion.model.vo.Promotion;
 
 public interface AdminDao {
 
-	List<ProductChildCategory> selectChildCategory(Map<String, Object> param);
 
-	int insertProduct(Product product);
-
-	int insertProducCategory(ProductCategory productCategory);
-	
-	int insertProductDetail(ProductDetail productDetail);
-	
-	List<ProductExtends> selectAllProductList(int offset, int limit);
-	
-	int selectTotalBoardCount();
 	
 	/**
 	 * 태영 끝-----------------------------------
