+++ src/main/java/com/kh/devrun/product/model/vo/ProductDetail.java
@@ -1,4 +1,4 @@
-package com.kh.devrun.product;
+package com.kh.devrun.product.model.vo;
 
 import java.io.Serializable;
