+++ src/main/java/com/kh/devrun/promotion/model/service/PromotionServiceImpl.java
@@ -11,7 +11,7 @@ import org.springframework.transaction.annotation.Isolation;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.kh.devrun.product.Product;
+import com.kh.devrun.product.model.vo.Product;
 import com.kh.devrun.promotion.model.dao.PromotionDao;
 import com.kh.devrun.promotion.model.exception.PromotionException;
 import com.kh.devrun.promotion.model.vo.Promotion;
